requirements.txt
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
idna==3.11
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.6.1
